chore(auth): move adapters to adapter layer
